+++ CHANGELOG.md
@@ -15,6 +15,7 @@ You can also check the [release page](https://github.com/visualize-admin/visuali
   - Temporal dimension X axis filtering in case of being a join by dimension
   - Sidebar content no longer gets unmounted when data is being re-fetched (e.g. when interacting with time slider in temporal X axes)
   - Reduced content layout shift when changing chart type
+  - Filters are now correctly applied when adding a new chart
 - Style
   - Introduced smaller UI improvements in the options panel